FTypes.Classes.FFunctor

Documentation

class FFunctor r where

Minimal complete definition

ffmap

Methods

ffmap :: (f ==> g) -> r f -> r g

(<<$) :: (forall a. g a) -> r f -> r g infixl 4

Replace all locations in the input with the same value. The default definition is fmap . const, but this may be overridden with a more efficient version.

(<<$>>) :: FFunctor r => (f ==> g) -> r f -> r g infixl 4